--- core/src/main/scala/org/apache/spark/mllib/optimization/OptimizerWithFrequency.scala
@@ -1,53 +1,0 @@
-/**
- *  Copyright (c) 2015 Intel Corporation 
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.spark.mllib.optimization
-
-import org.apache.spark.annotation.DeveloperApi
-import org.apache.spark.mllib.linalg.Vector
-import org.apache.spark.rdd.RDD
-
-/**
- * :: DeveloperApi ::
- * Trait for optimization problem solvers.
- *
- * Extension of MlLib's optimization trait that supports a frequency column.
- */
-@DeveloperApi
-trait OptimizerWithFrequency extends Serializable {
-
-  /**
-   * Solve the provided convex optimization problem.
-   */
-  def optimizeWithFrequency(data: RDD[(Double, Vector, Double)], initialWeights: Vector): Vector
-}
