+++ core/src/main/scala/org/trustedanalytics/sparktk/frame/internal/rdd/FrameRdd.scala
@@ -1,6 +1,9 @@
 package org.trustedanalytics.sparktk.frame.internal.rdd
 
 import breeze.linalg.DenseVector
+import org.apache.spark.mllib.regression.LabeledPoint
+import org.apache.spark.mllib.regression.org.trustedanalytics.sparktk.{ LabeledPointWithFrequency => SparktkLabeledPointWithFrequency }
+
 //import FrameOrderingUtils
 import org.trustedanalytics.sparktk.frame.DataTypes.{ float32, float64, int32, int64 }
 import org.trustedanalytics.sparktk.frame._
@@ -17,13 +20,12 @@ import org.apache.spark.mllib.stat.{ MultivariateStatisticalSummary, Statistics 
 //import org.apache.spark.atk.graph.{ EdgeWrapper, VertexWrapper }
 //import org.apache.spark.frame.ordering.FrameOrderingUtils
 import org.apache.spark.mllib.linalg.distributed.IndexedRow
-import org.apache.spark.mllib.linalg.{ Vector, Vectors, DenseVector => MllibDenseVector }
+import org.apache.spark.mllib.linalg.{ Vector, Vectors, VectorUDT, DenseVector => MllibDenseVector }
 import org.apache.spark.rdd.RDD
 import org.apache.spark.sql.catalyst.expressions.GenericRow
 import org.apache.spark.sql.types.{ ArrayType, BooleanType, ByteType, DateType, DecimalType, DoubleType, FloatType, IntegerType, LongType, ShortType, StringType, StructField, StructType, TimestampType }
 import org.apache.spark.sql.{ DataFrame, Row, SQLContext }
 import org.apache.spark.{ Partition, TaskContext }
-import org.apache.spark.mllib.regression.LabeledPoint
 //import org.trustedanalytics.atk.domain.schema.Column
 //import org.trustedanalytics.atk.domain.schema.DataTypes
 //import org.trustedanalytics.atk.domain.schema.DataTypes.DataType
@@ -121,6 +123,7 @@ class FrameRdd(val frameSchema: Schema, val prev: RDD[Row])
 
   /**
    * Convert FrameRdd into RDD[Vector] format required by MLLib
+   *
    * @param featureColumnNames Names of the frame's column(s) to be used
    * @return RDD of (org.apache.spark.mllib)Vector
    */
@@ -141,6 +144,7 @@ class FrameRdd(val frameSchema: Schema, val prev: RDD[Row])
 
   /**
    * Compute MLLib's MultivariateStatisticalSummary from FrameRdd
+   *
    * @param columnNames Names of the frame's column(s) whose column statistics are to be computed
    * @return MLLib's MultivariateStatisticalSummary
    */
@@ -151,6 +155,7 @@ class FrameRdd(val frameSchema: Schema, val prev: RDD[Row])
 
   /**
    * Convert FrameRdd to RDD[Vector] by mean centering the specified columns
+   *
    * @param featureColumnNames Names of the frame's column(s) to be used
    * @return RDD of (org.apache.spark.mllib)Vector
    */
@@ -164,6 +169,7 @@ class FrameRdd(val frameSchema: Schema, val prev: RDD[Row])
 
   /**
    * Convert FrameRdd to RDD[Vector]
+   *
    * @param featureColumnNames Names of the frame's column(s) to be used
    * @param columnWeights The weights of the columns
    * @return RDD of (org.apache.spark.mllib)Vector
@@ -181,6 +187,7 @@ class FrameRdd(val frameSchema: Schema, val prev: RDD[Row])
 
   /**
    * Convert FrameRdd to RDD[(Long, Long, Double)]
+   *
    * @param sourceColumnName Name of the frame's column storing the source id of the edge
    * @param destinationColumnName Name of the frame's column storing the destination id of the edge
    * @param edgeSimilarityColumnName Name of the frame's column storing the similarity between the source and destination
@@ -244,6 +251,7 @@ class FrameRdd(val frameSchema: Schema, val prev: RDD[Row])
 
   /**
    * Create a new FrameRdd that is only a subset of the columns of this FrameRdd
+   *
    * @param columnNames names to include
    * @return the FrameRdd with only some columns
    */
@@ -256,6 +264,7 @@ class FrameRdd(val frameSchema: Schema, val prev: RDD[Row])
 
   /**
    * Select a subset of columns while renaming them
+   *
    * @param columnNamesWithRename map of old names to new names
    * @return the new FrameRdd
    */
@@ -361,6 +370,7 @@ class FrameRdd(val frameSchema: Schema, val prev: RDD[Row])
 
   /**
    * Sort by one or more columns
+   *
    * @param columnNamesAndAscending column names to sort by, true for ascending, false for descending
    * @return the sorted Frame
    */
@@ -415,6 +425,7 @@ class FrameRdd(val frameSchema: Schema, val prev: RDD[Row])
 
   /**
    * Save this RDD to disk or other store
+   *
    * @param absolutePath location to store
    * @param storageFormat "file/parquet", "file/sequence", etc.
    */
@@ -478,6 +489,60 @@ class FrameRdd(val frameSchema: Schema, val prev: RDD[Row])
     new FrameRdd(frameSchema.addColumn(column), rows)
   }
 
+  /**
+   * Convert FrameRdd into RDD[LabeledPoint] format required by MLLib
+   */
+  def toLabeledPointRDD(labelColumnName: String, featureColumnNames: List[String]): RDD[LabeledPoint] = {
+    this.mapRows(row => {
+      val features = row.values(featureColumnNames).map(value => DataTypes.toDouble(value))
+      new LabeledPoint(DataTypes.toDouble(row.value(labelColumnName)), new MllibDenseVector(features.toArray))
+    })
+  }
+
+  /**
+   * Convert FrameRdd into RDD[LabeledPointWithFrequency] format required for updates in MLLib code
+   */
+  def toLabeledPointRDDWithFrequency(labelColumnName: String,
+                                     featureColumnNames: List[String],
+                                     frequencyColumnName: Option[String]): RDD[SparktkLabeledPointWithFrequency] = {
+    this.mapRows(row => {
+      val features = row.values(featureColumnNames).map(value => DataTypes.toDouble(value))
+      frequencyColumnName match {
+        case Some(freqColumn) => {
+          new SparktkLabeledPointWithFrequency(DataTypes.toDouble(row.value(labelColumnName)),
+            new MllibDenseVector(features.toArray), DataTypes.toDouble(row.value(freqColumn)))
+        }
+        case _ => {
+          new SparktkLabeledPointWithFrequency(DataTypes.toDouble(row.value(labelColumnName)),
+            new MllibDenseVector(features.toArray), DataTypes.toDouble(1.0))
+        }
+      }
+    })
+  }
+
+  /**
+   * Convert FrameRdd into labeled DataFrame with label of type double, and features of type vector
+   */
+  def toLabeledDataFrame(labelColumnName: String, featureColumnNames: List[String]): DataFrame = {
+    val labeledPointRdd = toLabeledPointRDD(labelColumnName, featureColumnNames)
+    val rowRdd: RDD[Row] = labeledPointRdd.map(labeledPoint => new GenericRow(Array[Any](labeledPoint.label, labeledPoint.features)))
+    val schema = StructType(Seq(StructField("label", DoubleType, true), StructField("features", new VectorUDT, true)))
+    new SQLContext(this.sparkContext).createDataFrame(rowRdd, schema)
+  }
+
+  /**
+   * Convert FrameRdd to DataFrame with features of type vector
+   */
+  def toLabeledDataFrame(featureColumnNames: List[String]): DataFrame = {
+    val vectorRdd: RDD[org.apache.spark.mllib.linalg.Vector] = this.mapRows(row => {
+      val features = row.values(featureColumnNames).map(value => DataTypes.toDouble(value))
+      new MllibDenseVector(features.toArray)
+    })
+    val rowRdd: RDD[Row] = vectorRdd.map(vector => new GenericRow(Array[Any](vector)))
+    val schema = StructType(Seq(StructField("features", new VectorUDT, true)))
+    new SQLContext(this.sparkContext).createDataFrame(rowRdd, schema)
+  }
+
 }
 
 /**
@@ -503,6 +568,7 @@ object FrameRdd {
 
   /**
    * converts a data frame to frame rdd
+   *
    * @param rdd a data frame
    * @return a frame rdd
    */
@@ -598,6 +664,7 @@ object FrameRdd {
 
   /**
    * Converts row object from an RDD[Array[Any]] to an RDD[Product] so that it can be used to create a SchemaRDD
+   *
    * @return RDD[org.apache.spark.sql.Row] with values equal to row object
    */
   def toRowRDD(schema: Schema, rows: RDD[Array[Any]]): RDD[org.apache.spark.sql.Row] = {
@@ -619,6 +686,7 @@ object FrameRdd {
 
   /**
    * Converts row object to RDD[IndexedRow] needed to create an IndexedRowMatrix
+   *
    * @param indexedRows Rows of the frame as RDD[Row]
    * @param frameSchema Schema of the frame
    * @param featureColumnNames List of the frame's column(s) to be used
@@ -636,6 +704,7 @@ object FrameRdd {
 
   /**
    * Converts row object to RDD[IndexedRow] needed to create an IndexedRowMatrix
+   *
    * @param indexedRows Rows of the frame as RDD[Row]
    * @param frameSchema Schema of the frame
    * @param featureColumnNames List of the frame's column(s) to be used
@@ -671,6 +740,7 @@ object FrameRdd {
 
   /**
    * Converts the schema object to a StructType for use in creating a SchemaRDD
+   *
    * @return StructType with StructFields corresponding to the columns of the schema object
    */
   def schemaToStructType(schema: Schema): StructType = {
@@ -696,6 +766,7 @@ object FrameRdd {
 
   /**
    * Converts the spark DataTypes to our schema Datatypes
+   *
    * @return our schema DataType
    */
   def sparkDataTypeToSchemaDataType(dataType: org.apache.spark.sql.types.DataType): DataTypes.DataType = {
@@ -730,6 +801,7 @@ object FrameRdd {
 
   /**
    * Converts a spark dataType (as string)to our schema Datatype
+   *
    * @param sparkDataType spark data type
    * @return a DataType
    */
@@ -750,6 +822,7 @@ object FrameRdd {
 
   /**
    * Converts the schema object to a StructType for use in creating a SchemaRDD
+   *
    * @return StructType with StructFields corresponding to the columns of the schema object
    */
   def schemaToAvroType(schema: Schema): Seq[(String, String)] = {
